+++ framework_v3/composer.json
@@ -1,0 +1,5 @@
+{
+    "require": {
+        "twig/twig": "^3.0"
+    }
+}
