+++ framework_v3/public/index.php
@@ -1,10 +1,13 @@
 <?php
 use App\services\Autoloader;
+use App\services\RenderTemplate;
 
 include dirname(__DIR__) . '/services/Autoloader.php';
 spl_autoload_register([(new Autoloader()), 'loadClass']);
+require_once dirname(__DIR__) . '/vendor/autoload.php';
 
-$controller = 'user';
+
+$controller = 'product';
 if ($_GET['c']) {
   $controller = $_GET['c'];
 }
@@ -18,7 +21,7 @@ $controllerName = 'App\\controllers\\' . ucfirst($controller) . 'Controller';
 
 if (class_exists($controllerName)) {
   /** @var \App\controllers\UserController $realController */
-  $realController = new $controllerName();
+  $realController = new $controllerName(new RenderTemplate());
   $content = $realController->run($action);
 
   if (!empty($content)) {
