+++ framework_v3/services/IRenderer.php
@@ -1,0 +1,10 @@
+<?php
+
+
+namespace App\services;
+
+
+interface IRenderer
+{
+  public function render($template, $params = []);
+}
